Fixes in Luby and Geometric cutoff
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@13504 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/search/cutoff.hpp
@@ -97,7 +97,7 @@ namespace Gecode { namespace Search {
     friend class Cutoff;
   private:
     /// Current cutoff value
-    unsigned long int n;
+    double n;
     /// Base
     double base;
     /// Constructor
